stable unfinished fix/checkpoint
