+++ package.json
@@ -18,6 +18,7 @@
     "@vercel/analytics": "^1.4.0",
     "@vercel/speed-insights": "^1.1.0",
     "gsap": "^3.12.5",
+    "imagesloaded": "^5.0.0",
     "lil-gui": "^0.19.2",
     "three": "^0.166.1"
   }
