+++ src/index.html
@@ -874,10 +874,8 @@ Date: 11-24
         <p>2024</p>
       </section>
     </main>
-    <script src="js/gsap.min.js"></script>
-    <script src="js/ScrollTrigger.min.js"></script>
-    <script src="js/imagesloaded.pkgd.min.js"></script>
-    <script type="module" src="js/menu.js"></script>
-    <script type="module" src="js/index.js"></script>
+    <!-- <script type="module" src="/js/gsap.min.js"></script> -->
+    <script type="module" src="/js/menu.js"></script>
+    <script type="module" src="/js/index.js"></script>
   </body>
 </html>
