+++ src/js/effect-1/stackMotionEffect1.js
@@ -1,10 +1,12 @@
-import { throttle } from '../main-utils.js';
+import { throttle } from '/js/main-utils.js';
+import gsap from 'gsap';
 
+// Keeps track of the window's size for responsive adjustments.
 let winsize = {width: window.innerWidth, height: window.innerHeight};
 
 export class StackMotionEffect {
   constructor(stackEl) {
-    // Check if the provided element is valid.
+    // Validates the input element to ensure it's an HTML element.
     if (!stackEl || !(stackEl instanceof HTMLElement)) {
       throw new Error('Invalid element provided.');
     }
@@ -14,7 +16,7 @@ export class StackMotionEffect {
     this.imageElements = [this.contentElement.querySelectorAll('.card')];
     this.imagesTotal = this.imageElements.length;
 
-    // Set up the effect for the provided element.
+    // Calls the method to set up the initial effect.
     this.initializeEffect(stackEl);
   }
   
@@ -22,7 +24,8 @@ export class StackMotionEffect {
   initializeEffect(element) {
     // Scroll effect.
     this.scroll();
-
+    
+    // Throttles resize event to optimize performance and re-calculate sizes and effect on resize.
     const throttledResize = throttle(() => {
       winsize = { width: window.innerWidth, height: window.innerHeight };
       this.scroll();
@@ -30,15 +33,19 @@ export class StackMotionEffect {
     window.addEventListener('resize', throttledResize);
   }
 
+  // Defines the scroll effect logic for the stack.
   scroll() {
-    // Let's set the initial rotation for the content element
-    this.contentElement.style.transform = 'rotate3d(1, 0, 0, 25deg) rotate3d(0, 1, 0, -50deg) rotate3d(0, 0, 1, 25deg)';
+    // Initially hides the content element and prepares it for the animation by setting its transform property. 
+    // This sets the initial 3D rotation of the stack and its cards, defining their starting visual appearance.
+    this.contentElement.style.transform = 'rotate3d(1, 0, 0, -25deg) rotate3d(0, 1, 0, 50deg) rotate3d(0, 0, 1, 25deg)';
     this.contentElement.style.opacity = 0;
 
+    // Clears previous timeline if exists to prevent conflicts.
     if (this.tl) {
       this.tl.kill();
     }
 
+    // Creates a new timeline for the scroll-triggered animation.
     this.tl = gsap.timeline({
       defaults: {
         ease: 'power1',
@@ -48,29 +55,33 @@ export class StackMotionEffect {
         start: 'top center',
         end: '+=150%',
         scrub: true,
+        // Sets opacity to 1 when the element comes into view.
         onEnter: () => gsap.set(this.contentElement, {opacity: 1}),
         onEnterBack: () => gsap.set(this.contentElement, {opacity: 1}),
+        // Hides the element when it leaves the view.
         onLeave: () => gsap.set(this.contentElement, {opacity: 0}),
-        onLeaveBack: () => gsap.set(this.contentElement, {opacity: 0}),
+        onLeaveBack: () => gsap.set(this.contentElement, {opacity: 0})
       },
     })
     .fromTo(this.imageElements, {
-      z: (pos) => -2.5 * winsize.width/2 - pos * 0.07 * winsize.width,
+      // Animates from a starting z position based on the window size.
+      z: (pos) => -2.65 * winsize.width - pos * 0.03 * winsize.width,
     }, {
-      z: (pos) => 2.5 * winsize.width + (this.imagesTotal - pos - 1) * 0.07 * winsize.width,
+      // Animates to an ending z position, creating a 3D effect as elements scroll.
+      z: (pos) => 1.4 * winsize.width + (this.imagesTotal - pos - 1) * 0.03 * winsize.width,
     }, 0)
     .fromTo(this.imageElements, {
-      rotationZ: 10,
+      rotationZ: -220,
     }, {
-      rotationX: 20,
-      rotationZ: 280,
-      yPercent: -100,
+      rotationY: -30,
+      rotationZ: 120,
+      // Stagger effect for individual elements to animate sequentially.
       stagger: 0.005,
     }, 0)
     /*.fromTo(this.imageElements, {
       filter: 'brightness(20%)',
     }, {
-      filter: 'brightness(350%)',
+      filter: 'brightness(150%)',
       stagger: 0.005,
     }, 0);*/
   }
