+++ src/js/main-utils.js
@@ -4,6 +4,9 @@
  * @param {string} [selector='img'] - CSS selector for target images.
  * @returns {Promise} - Resolves when all specified images are loaded.
  */
+
+import imagesLoaded from "imagesloaded";
+
 export const preloadImages = (selector = 'img') => {
     return new Promise((resolve) => {
         // The imagesLoaded library is used to ensure all images (including backgrounds) are fully loaded.
